# Add load_file performance benchmark tooling