--- a/tests/benchmark/config-sample.json
+++ b/tests/benchmark/config-sample.json
@@ -1,0 +1,53 @@
+{
+  "databases": [
+    {
+      "name": "postgres",
+      "params": {
+        "database": "postgres",
+        "conn_id": "postgres_conn",
+        "schema": "postgres"
+      }
+    },
+    {
+      "name": "snowflake",
+      "params": {
+        "database": "DWH_LEGACY",
+        "conn_id": "snowflake_conn",
+        "schema": "SANDBOX_TATIANAALCHUEYR"
+      }
+    },
+    {
+      "name": "bigquery",
+      "params": {
+        "database": "bigquery",
+        "conn_id": "bigquery"
+      }
+    }
+  ],
+  "datasets": [
+    {
+        "name": "few_kb",
+        "size": "45K",
+        "path":"/tmp/covid_overview.csv",
+        "rows": 732
+    },
+    {
+        "name": "many_kb",
+        "size": "472K",
+        "path": "/tmp/artist_data.csv",
+        "rows": 3532
+    },
+    {
+        "name": "few_mb",
+        "size": "20M",
+        "path": "/tmp/title_ratings.csv",
+        "rows": 1213991
+    },
+    {
+        "name": "many_mb",
+        "size": "279M",
+        "path": "/tmp/github_timeline.csv",
+        "rows": 385817
+    }
+  ]
+}