--- a/tests/benchmark/config.json
+++ b/tests/benchmark/config.json
@@ -1,0 +1,26 @@
+{
+  "databases": [
+    {
+      "name": "postgres",
+      "params": {
+        "database": "postgres",
+        "conn_id": "postgres_conn",
+        "schema": "postgres"
+      }
+    }
+  ],
+  "datasets": [
+    {
+        "name": "few_kb",
+        "size": "45K",
+        "path":"/home/tati/Data/covid_overview.csv",
+        "rows": 732
+    },
+    {
+        "name": "many_kb",
+        "size": "472K",
+        "path": "/home/tati/Data/artist_data.csv",
+        "rows": 3532
+    }
+  ]
+}