As part of this PR, I implemented a benchmark tool to support us in analyzing the performance of Astro operators and tasks.
At the moment it is being used to evaluate `load_file`, but in the next PR, it will be generalized to be used with other operators.

Relates to issue #108 

Considering the chunk sizes are set to 2,000, the performance of the operators was:

### Database: BigQuery

| database   | dataset   | total_time   | memory_rss   | memory_shared   | memory_pss   | cpu_time_user   | cpu_time_system   |
|:-----------|:----------|:-------------|:-------------|:----------------|:-------------|:----------------|:------------------|
| bigquery   | few_kb    | 11.28s       | 120.73MB     | 40.91MB         | 118.02MB     | 3.5s            | 1.63s             |
| bigquery   | many_kb   | 18.57s       | 122.39MB     | 40.68MB         | 119.74MB     | 3.55s           | 1.58s             |
| bigquery   | few_mb    | 42.99min     | 140.99MB     | 40.94MB         | 138.21MB     | 20.07s          | 2.27s             |



### Database: postgres

| database   | dataset   | total_time   | memory_rss   | memory_shared   | memory_pss   | cpu_time_user   | cpu_time_system   |
|:-----------|:----------|:-------------|:-------------|:----------------|:-------------|:----------------|:------------------|
| postgres   | few_kb    | 1.73s        | 125.62MB     | 41.89MB         | 122.95MB     | 1.95s           | 1.4s              |
| postgres   | many_kb   | 2.24s        | 138.5MB      | 41.78MB         | 135.95MB     | 2.4s            | 1.38s             |
| postgres   | few_mb    | 1.23min      | 156.85MB     | 41.96MB         | 154.15MB     | 1.09min         | 1.74s             |
| postgres   | many_mb   | 23.94min     | 1.23GB       | 41.95MB         | 1.23GB       | 22.15min        | 31.5s            |


### Database: snowflake

| database   | dataset   | total_time   | memory_rss   | memory_shared   | memory_pss   | cpu_time_user   | cpu_time_system   |
|:-----------|:----------|:-------------|:-------------|:----------------|:-------------|:----------------|:------------------|
| snowflake  | few_kb    | 14.28s       | 136.06MB     | 50.39MB         | 132.96MB     | 3.12s           | 1.48s             |
| snowflake  | many_kb   | 12.42s       | 138.83MB     | 50.22MB         | 135.85MB     | 3.17s           | 1.48s             |
| snowflake  | few_mb    | 7.21min      | 264.67MB     | 50.2MB          | 261.61MB     | 25.95s          | 3.27s             |
| snowflake  | many_mb   | 4.86min      | 975.47MB     | 50.48MB         | 972.4MB      | 41.67s          | 5.03s             |